refactor: put some SuperNodes logic in the SuperNodes repo
